Fixed #1158 - 2.0: getLong/setLong does not work as expected
- Casting long value to int causes data loss. Main fix is in litecore
- Update unit tests
- Replaced HashMap with TreeMap in Unit Test because TreeMap can guarantee key order.
- Make Dictionary.toMap() uses TreeMap as implementation instead of HashMap.

+++ shared/src/main/java/com/couchbase/lite/ReadOnlyArray.java
@@ -88,7 +88,7 @@ public class ReadOnlyArray implements ReadOnlyArrayInterface, FleeceEncodable {
      */
     @Override
     public int getInt(int index) {
-        return fleeceValue(index).asInt();
+        return (int)fleeceValue(index).asInt();
     }
 
     /**
